MASShortcut.xcodeproj: Change settings to work in Low Battery Yup
Change a number of settings to work in the context of my application and
my computer/development environment.

* Set build architecture to Standard (32/64-bit Intel)
* Set deployment target to 10.8
* Link MASShortcut.framework with Carbon.framework (fix
* `CopySymbolicHotKeys` etc. undeclared error)
* Delete test target because it was failing on:

		target specifies product type 'com.apple.product-type.bundle.unit-test', but there's no such product type for the 'macosx' platform

Since I don't appear to have that product type in my Xcode 4.5.2, I
figured I wasn't really going to run the tests anyway so let's just
delete the target in order to get the framework to build.